+++ src/utility/ethernetif.h
@@ -47,14 +47,11 @@
 
 #ifndef __ETHERNETIF_H__
 #define __ETHERNETIF_H__
-
-#include "lwip/err.h"
-#include "lwip/netif.h"
-
 #ifdef __cplusplus
  extern "C" {
 #endif
-
+#include "lwip/err.h"
+#include "lwip/netif.h"
 /* Exported types ------------------------------------------------------------*/
 err_t ethernetif_init(struct netif *netif);
 void ethernetif_input(struct netif *netif);
@@ -63,6 +60,9 @@ void ethernetif_update_config(struct netif *netif);
 void ethernetif_notify_conn_changed(struct netif *netif);
 
 void ethernetif_set_mac_addr(const uint8_t *mac);
+
+err_t igmp_mac_filter( struct netif *netif, const ip4_addr_t *ip4_addr, u8_t action );
+void register_multicast_address(const uint8_t *mac);
 
 #ifdef __cplusplus
 }
