+++ src/utility/stm32_eth.cpp
@@ -47,9 +47,9 @@
 #include "lwip/prot/dhcp.h"
 #include "lwip/dns.h"
 
-#ifdef __cplusplus
- extern "C" {
-#endif
+// #ifdef __cplusplus
+//  extern "C" {
+// #endif
 
 /* Check ethernet link status every seconds */
 #define TIME_CHECK_ETH_LINK_STATE 500U
@@ -456,8 +456,6 @@ void ethernetif_notify_conn_changed(struct netif *netif)
 {
   if(netif_is_link_up(netif))
   {
-    printf("Link up\n");
-
     /* Update DHCP state machine if DHCP used */
     if(DHCP_Started_by_user == 1) {
       DHCP_state = DHCP_START;
@@ -475,8 +473,6 @@ void ethernetif_notify_conn_changed(struct netif *netif)
 
     /*  When the netif link is down this function must be called.*/
     netif_set_down(netif);
-
-    printf("Link down\n");
   }
 }
 
@@ -761,6 +757,10 @@ void udp_receive_callback(void *arg, struct udp_pcb *pcb, struct pbuf *p,
 
     ip_addr_copy(udp_arg->ip, *addr);
     udp_arg->port = port;
+
+    if(udp_arg->onDataArrival != NULL){
+      udp_arg->onDataArrival();
+    }
   } else {
     pbuf_free(p);
   }
@@ -1013,6 +1013,6 @@ void tcp_connection_close(struct tcp_pcb *tpcb, struct tcp_struct *tcp)
 
 #endif /* LWIP_TCP */
 
-#ifdef __cplusplus
-}
-#endif
+// #ifdef __cplusplus
+// }
+// #endif
